bot/requirements.txt
aiogram
python-dotenv
paramiko
sqlalchemy[asyncio]
asyncpg
